cleanup(ikigai): drop 7 preview routes, rename v4 → v1, remove presenter button
Frank's directive: 'the normal version is best, just keep v4 rename to v1
keep /workshops/ikigai-branding/present but remove /workshops/ikigai/v1-9
not good. also remove the button with presenter mode not good presenting mode'

What shipped:

1. Canonical cleanup (app/workshops/ikigai-branding/page.tsx)
   - Dropped the hero footer 'Facilitating? Open presenter mode' link
   - Dropped the Presentation lucide-react icon import (unused)
   - Dropped the PresenterOverlay component + ?present=1 useEffect +
     presenterActive useState + PRESENTER_SECTIONS const + useState/useEffect
     React imports (all dead now)
   - Updated comment header about preview archives (V1 only now)
   - Canonical content otherwise unchanged — Frank: 'the normal version is best'

2. v4 → v1 rename + archive framing
   - Moved app/workshops/ikigai/v4/{page,layout}.tsx → v1/{page,layout}.tsx
   - Fixed v1/page.tsx outbound dead links (was pointing at /v2 + /v3 which
     get deleted in this same commit). Replaced 3-link footer nav with one
     'Back to the workshop' link + 'V1 archive' italic label.
   - Updated v1/layout.tsx — robots noindex, title says 'V1 Archive',
     description points readers to canonical workshop.

3. Deleted 7 preview folders (~5,000 lines)
   - /v2 (page.tsx)
   - /v3 (page.tsx + layout.tsx)
   - /v4 (page.tsx + layout.tsx) — content lives at /v1 now
   - /v5 (page.tsx + layout.tsx)
   - /v6 (page.tsx + layout.tsx)
   - /v7 (page.tsx + layout.tsx)
   - /v8 (page.tsx + layout.tsx)
   - /v9 (page.tsx)

4. Research page cleanup (app/research/blue-zones-ikigai-ai-era/page.tsx)
   - Deleted the 'Design comparisons' section that showcased V2/V3/V4 cards
     (~80 lines). Section's premise — 'four versions for design comparison' —
     no longer holds after the deletes.
   - Replaced with a single 'Open the workshop' CTA pointing at canonical.
   - Repointed 2 stray inline /workshops/ikigai/v4 links → canonical.

5. Workshop-day handover doc updated
   - Removed V9 Solo Edition row from URL table
   - Added V1 archive note for facilitator reference
   - Updated prompt-library description (no count claim)

The /workshops/ikigai-branding/present route is untouched — facilitator
mode still available, just no longer surfaced from canonical UI.
